[WICS-3] Added a new repository for ConceptType and Added Javadocs
