+++ wic/src/main/java/edu/asu/diging/wic/web/PersonController.java
@@ -23,11 +23,11 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 import edu.asu.diging.wic.core.conceptpower.IConceptpowerCache;
 import edu.asu.diging.wic.core.graphs.IGraphDBConnection;
-import edu.asu.diging.wic.core.graphs.repository.GraphRepository;
 import edu.asu.diging.wic.core.model.IConcept;
 import edu.asu.diging.wic.core.model.impl.ConceptText;
 import edu.asu.diging.wic.core.model.impl.Edge;
 import edu.asu.diging.wic.core.model.impl.Node;
+import edu.asu.diging.wic.core.repository.GraphRepository;
 import edu.asu.diging.wic.core.service.IConceptTextService;
 import edu.asu.diging.wic.web.cytoscape.Data;
 import edu.asu.diging.wic.web.cytoscape.EdgeData;
